# src/main/resources/openapi/v1/product-openapi.yaml
openapi: 3.1.0
info:
  title: Product Catalog API
  description: |
    REST API for managing products in an e-commerce platform.
    
    Supports:
    - CRUD operations on products
    - Paginated & filtered product search
    - Quick price lookup
    
    Most endpoints require authentication (usually admin role for create/update/delete).
    Search and price lookup are typically public or customer-accessible.
  version: 1.0.0
  contact:
    name: Infinity Tech Support
    email: infinitytech@support.co.za
  license:
    name: Proprietary License - All rights reserved.
    url: https://infinitytech.co.za/terms

servers:
  - url: https://api.infinitytech.com/v1
    description: Production server
  - url: https://staging-api.infinitytech.com/v1
    description: Staging server

tags:
  - name: Products
    description: Operations for creating, reading, updating, deleting and searching products

paths:
  /api/v1/products/create:
    post:
      tags:
        - Products
      summary: Create a new product
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDto'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
              example:
                statusCode: 201
                message: "Product created successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  
  /api/v1/products/update/{productId}:
    patch:
      tags:
        - Products
      summary: Update an existing product
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDto'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Product not found
  
  /api/v1/products/find/{productId}:
    get:
      tags:
        - Products
      summary: Get product by ID
      operationId: findProductById
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '404':
          description: Product not found
  
  /api/v1/products/delete/{productId}:
    delete:
      tags:
        - Products
      summary: Delete a product
      operationId: deleteProduct
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Product deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
              example:
                statusCode: 200
                message: "Operation successful"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Product not found
  
  /api/v1/products/all:
    get:
      tags:
        - Products
      summary: Get all products (unpaginated)
      description: Returns all products in the catalog. Use `/search` for paginated/filtered results.
      operationId: getAllProducts
      responses:
        '200':
          description: List of all products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDto'
  
  /api/v1/products/search:
    get:
      tags:
        - Products
      summary: Search and filter products with pagination and sorting
      description: |
        Advanced search endpoint supporting name search, price range, pagination and sorting.
        Example: `/search?name=laptop&minPrice=5000&maxPrice=25000&page=0&size=12&sortField=price&sortDir=desc`
      operationId: searchProducts
      parameters:
        - name: productName
          in: query
          required: false
          schema:
            type: string
          description: Partial or full product name
        - name: minPrice
          in: query
          required: false
          schema:
            type: number
            format: decimal
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
            format: decimal
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sortField
          in: query
          required: false
          schema:
            type: string
            default: productId
            enum: [productId, name, price, createdAt]
        - name: sortDir
          in: query
          required: false
          schema:
            type: string
            default: asc
            enum: [asc, desc, ASC, DESC]
      responses:
        '200':
          description: Paginated list of matching products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDto'
        '400':
          $ref: '#/components/responses/BadRequest'
  
  /api/v1/products/{productId}/price:
    get:
      tags:
        - Products
      summary: Get current price of a product
      operationId: getProductPrice
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Current product price
          content:
            application/json:
              schema:
                type: number
                format: decimal
              example: 12499.99
        '404':
          description: Product not found

components:
  schemas:
    
    ProductDto:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
          minLength: 3
          maxLength: 200
          example: "MacBook Pro 16\" M3 Max"
        description:
          type: string
          example: "Apple MacBook Pro with M3 Max chip, 48GB RAM, 1TB SSD"
        price:
          type: number
          format: decimal
          minimum: 0
          example: 124999.99
        currency:
          type: string
          default: ZAR
          example: ZAR
        stockQuantity:
          type: integer
          minimum: 0
          example: 42
        sku:
          type: string
          example: MBP16-M3MAX-1TB-SILVER
        category:
          type: string
          example: "Laptops"
        brand:
          type: string
          example: "Apple"
        images:
          type: array
          items:
            type: string
            format: uri
        active:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required:
        - name
        - price
        - stockQuantity
    
    ResponseDto:
      type: object
      properties:
        statusCode:
          type: integer
          example: 201
        message:
          type: string
          example: "Product created successfully"
      required:
        - statusCode
        - message
    
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions (e.g. not admin)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT